+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2020-10-12
+
+### Added
+- Added a transparency scaler.
+- Added settings to control the Indexers thermal and performance actions when using the Device Simulator.
+- Updated the version defines for the device simulator to support it in 2021.1 without package.
+
+### Changed
+- Session bugfixes for Adaptive Performance provider subsystem management.
+
 ## [2.0.2] - 2019-08-21
 
 ### Changed